Finished 4.21

+++ Part 4/blog list/controllers/blogs.js
@@ -54,8 +54,23 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    const blog = await Blog.findById(request.params.id)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if(!decodedToken.id){
+        return response.status(401).json({ Error : 'Missing or Invalid token' })
+    }
+    if(blog.user.toString() === decodedToken.id.toString()){
+        await Blog.findByIdAndDelete(request.params.id)
+        response.status(204).end()
+    }else{
+        response.status(401).json({ Error : 'Do not Have permission to delete this blog' })
+    }
+
+    console.log(blog.user.toString(), decodedToken.id.toString())
+
+    // // await Blog.findByIdAndDelete(request.params.id)
+    // response.status(204).end()
 })
 
 blogsRouter.put('/:id', async (request, response) => {
